requirements.txt
# discord
discord.py

# data
spotipy
yt-dlp

# music
lyricsgenius
PyNaCl

# system
python-dotenv